requirements.txt
Django>=3.2,<4
djangorestframework
celery
psycopg2-binary
